Revert "Moved PROJECTION_*** constants from /scene to /core/shape (#4735)"
This reverts commit 8b93e8211bf92c701e8cc902ec540827747fe373.
